Add MPRIS background blur for app icons (#369)
* Fix default mpris icon size

The widget was was using the .ui size instead of the user config
size in the following situations:

- there was no media player app
- the media player app icon was used
- the audio-x-generic-symbolic icon was used

This fix sets the default size for the album art icon according to the
user's config.

set_from_gicon and set_from_icon_name receive a Gtk.IconSize enum value,
not an int, so we need to call set_pixel_size to set the proper size.

* Add MPRIS background blur for app icons

---------

Co-authored-by: Erik Reider <35975961+ErikReider@users.noreply.github.com>