Consolidate --noti-bg and --noti-bg-alpha in a single variable
The values are put together in --noti-bg-normal. In places where the
alpha value differs, GTK 3.0's alpha() function is used to change it.

Since SASS also has its own alpha() function unrelated to GTK's, it
causes issues in compilation. Implement a custom function which
generates the function properly in the final CSS.

They are now indirectly used and not removed for backwards compatibility.